feat: 공통 MetadataPickerModal — LoRA/Model picker 단일화 — #260 Phase 3
검색/필터/페이지네이션/sync/NSFW 토글까지 포함한 picker modal 본체를
공통화. LoraListModal + ModelPickerGrid 두 파일을 모두 제거하고 단일
`MetadataPickerModal` 사용.

## 신규 컴포넌트
`frontend/src/components/common/MetadataPickerModal.js` (~370 라인)

**Props**:
- `kind`: 'lora' | 'model'
- `open` / `onClose` / `serverId` / `workboardId`(LoRA fallback) / `isAdmin`
- `mode`: 'select-single' | 'multi-add' | 'prompt-insert'
- `selectedItem` (filename), `onPrimary` (rawItem)
- `onTrainedWordClick` (LoRA prompt-insert 모드)
- `allowedModelTypes` (Model picker 작업판 제약 #252)
- `title` (옵션 — 기본 "<label> 선택")

**Per-kind adapter** (`KIND_ADAPTERS` 객체로 캡슐화):
- `fetch` — kind 별 API (serverAPI.getLoras / getDetailedModels), workboardId fallback (LoRA)
- `extractList` / `extractPagination` / `extractCacheInfo` / `extractAvailableBaseModels` — 응답 필드명 차이 흡수
- `sync` / `getStatus` — 동기화
- `normalize` — Phase 1 helpers
- `nsfwItemPreference` — NSFW item 자체 숨김 preference key (LoRA 만)

**기능**:
- 검색 (300ms debounce) + baseModel filter (allowedModelTypes 와 교집합)
- 페이지네이션 + sync 트리거 + 진행률 LinearProgress + 상태 폴링 (2초)
- NSFW 미리보기 이미지 토글
- NSFW item 자체 토글 (LoRA 만 — userPreferences.nsfwLoraFilter)
- 작업판 허용 타입 chip 안내 (Model 만)
- 카드 그리드 (MetadataItemCard 사용)

## 사용처 교체
**`ImageGeneration.js`**:
- LoraListModal / ModelPickerGrid import 제거 → MetadataPickerModal
- LoRA 의 prompt-insert 핸들러 (handleAddLoraToPrompt, handleInsertTriggerWord)
  를 ImageGeneration 으로 이동 (기존 LoraListModal 내부 로직 그대로)
- LoRA picker: kind="lora" mode="prompt-insert"
- Model picker: kind="model" mode="select-single"

## 삭제
- `LoraListModal.js` (544 라인)
- `ModelPickerGrid.js` (379 라인)

## 효과
- 두 picker 의 유지보수 표면이 한 컴포넌트로 통합 (~370 라인)
- 새 기능 (검색 개선, sync UI 개선 등) 추가 시 한 곳만 수정
- LoRA 와 Model 의 사용자 경험 일관성 (확장 카드, 트리거 워드, NSFW 필터,
  sync 진행률 등 모두 동일하게 동작)
- 코드 변동: -932 / +62 (순감소 -870 라인)

## 회귀 영역 검토
- LoRA prompt-insert: 핸들러를 ImageGeneration 으로 이동했지만 로직 동일.
  insertLoraTag / insertTriggerWordWithLora promptUtils 사용 그대로
- LoRA workboardId fallback: adapter.fetch 가 serverId 없을 때 workboardAPI 호출
- Model 단일 선택 + cardClickable: 카드 클릭 시 onPrimary + onClose
- allowedModelTypes: model adapter 가 query param 으로 backend 전달 (Phase 3 #252)

## 후속
- Phase 4: 공통 admin 페이지 + LoRA 즐겨찾기 모델에도 적용

Refs #260
